automator/automation_engine.py
# automator/automation_engine.py
import subprocess, time, os, logging
import pyautogui, pytesseract
from PIL import Image

class AutomationEngine:
    def __init__(self): pass

    def run(self, profile):
        exe = profile.get('game_path')
        try:
            if exe:
                logging.info(f'Starting game: {exe}')
                subprocess.Popen([exe], shell=True)
                time.sleep(10)
                self._perform_actions(profile)
                logging.info('Automation completed successfully.')
            else:
                logging.error('No game path in profile.')
        except Exception as e:
            logging.error(f'Automation failed: {e}')

    def _perform_actions(self, profile):
        logging.info('Performing automation actions...')
        screenshot = pyautogui.screenshot()
        text = pytesseract.image_to_string(screenshot)
        logging.info(f'OCR read: {text.strip()}')
        # Add game-specific pyautogui steps here
