requirements.txt
pyautogui
pillow
schedule
pytesseract
